requirements.txt
django==5.1
psycopg2-binary
djangorestframework
djangorestframework-simplejwt
django-cors-headers
python-dateutil